Title: 컴퓨터와 함께하는 숫자 게임 2, Time: 129ms, Memory: 10MB, Status: Runtime Error - Codetree

+++ 250603/컴퓨터와 함께하는 숫자 게임 2/play-number-game-with-computer-2.java
@@ -31,7 +31,7 @@ public class Main {
         long a = sc.nextInt();
         long b = sc.nextInt();
         long mmax = Long.MIN_VALUE;
-        int minx = Long.MAX_VALUE;
+        long minx = Long.MAX_VALUE;
         for (long i = a; i <= b; i++) {
 
             int cnt = find(i);
